refactor(folder structure): reorganize folder structure for better modularity
