+++ .eslintrc
@@ -49,7 +49,7 @@
           ["@assets", "./src/assets"],
           ["@hooks", "./src/hooks"],
           ["@pages", "./src/pages"],
-          ["@data", "./src/data"],
+          ["@mocks", "./src/mocks"],
           ["@routes", "./src/routes"],
           ["@utils", "./src/utils"],
           ["@context", "./src/context"]
