+++ src/components/common/index.ts
@@ -1,13 +1,12 @@
 export { default as Button } from './Button';
-export { default as Footer } from './Footer';
+export { default as Footer } from '../features/Layout/Footer';
 export { default as InputField } from './Form/InputField';
 export { default as Image } from './Image';
 export { default as GoodsItem } from './GoodsItem';
 export { default as Ranking } from './GoodsItem/Ranking';
-export { default as Header } from './Header';
+export { default as Header } from '../features/Layout/Header';
 export { default as Container } from './Layout/Container';
 export { default as Grid } from './Layout/Grid';
 export { default as Inner } from './Layout/Inner';
-export { default as PageWrapper } from './Layout/PageWrapper';
 export { default as SectionHeader } from './SectionHeader';
-export { default as PrivateRoute } from './PrivateRoute';
+export { default as PrivateRoute } from '../../routes/components/PrivateRoute';
