+++ src/components/features/Home/FriendGiftRecommendation/index.stories.ts
