+++ src/components/features/Home/FriendSelector/index.tsx
