--- src/pages/MyAccount.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-import { PageWrapper } from '@components/common';
-import { AccountOverview } from '@components/account';
-
-export default function MyAccount() {
-  return (
-    <PageWrapper>
-      <AccountOverview />
-    </PageWrapper>
-  );
-}
