+++ src/pages/MyAccount/index.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import Layout from '@components/features/Layout';
+import { AccountOverview } from '@components/features/Account';
+
+export default function MyAccount() {
+  return (
+    <Layout>
+      <AccountOverview />
+    </Layout>
+  );
+}
