+++ src/routes/components/PrivateRoute/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { useAuth } from '@context/auth/useAuth';
-import { ROUTE_PATH } from '@routes/constants';
+import { ROUTE_PATH } from '@routes/path';
 import { Navigate, Outlet } from 'react-router-dom';
 
 export default function PrivateRoute() {
